bug fix to age_DNA
found bug where function was only doing deamination on the first 77% of rows. Caused by indexing error (i.e. not grabbing the indices of the transition sites)

updated tests for random_impute. Added comment to clarify that r refers to the order of the NA, rather than the row index.